pyproject.toml
[project]
name = "google-calender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.3.3",
    "google-api-python-client>=2.169.0",
    "google-auth>=2.40.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "openai-agents[litellm]>=0.0.14",
    "python-dateutil>=2.9.0.post0",
]
